refactor: Move files to right place
